Update Documentation for Icon Menus
