--- docs/src/app/components/raw-code/icon-menus-code.txt
@@ -1,14 +1,0 @@
-//We're working on migrating some of our components to use a new implementation of menus.
-//If you'd like to use the new menu before our migration is complete, please directly
-//require them like this:
-
-import MenuItem from 'material-ui/lib/menus/menu-item';
-import Divider from 'material-ui/lib/divider';
-
-<IconMenu iconButtonElement={iconButtonElement}>
-  <MenuItem primaryText="Refresh" />
-  <MenuItem primaryText="Send feedback" />
-  <MenuItem primaryText="Settings" />
-  <MenuItem primaryText="Help" />
-  <MenuItem primaryText="Sign out" />
-</IconMenu>
